refactor: share labels URL resolution
Centralize server-side labels base URL resolution so the SDK config, labels proxy, and metadata logo URLs use the same environment fallback.
